General tidy up of project
There were 2 sets of .ino and header files in the repo, I have made sure there is only 1 set of the most up to date files and renamed the folder they reside in appropriately. I also added the library sources for external libraries.

I uploaded the brd and sch files as well as supporting files from the tetra repo. I also added a full set of Gerbers ready for manufacture.

On the .brd file the labels were the wrong way around for the resistor and the LED so I fixed that.

I removed a workspace file that was surplus to requirements.

The reason I have chosen to fork Patricks original repo as opposed to the work in the tetra repo is because it will be easier to maintain this repo as a self contained piece of work rather than working within the tetra repo.

The info in the tetra repo is incorrect as it stands as I believe the wrong .ino file has been uploaded.